kafka_2.13-4.1.0/requirements.txt
kafka-python
aiosqlite